feat: add the backend kind to improve the log
